backend/requirements.txt
streamlit>=1.40.0
fastapi>=0.112.4
langchain>=0.3.3
langchain-community>=0.3.2
langchain-core>=0.3.10
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-pinecone>=0.2.0
langgraph>=0.2.38
pinecone-client>=5.0.1
transformers>=4.44.2
python-dotenv>=1.0.1
numpy>=1.26.4
pandas>=2.2.3
nltk>=3.9.1
scikit-learn>=1.5.2
sentence-transformers>=3.1.0
uvicorn>=0.25.0
azure-storage-blob>=12.23.1
python-multipart>=0.0.9
pydantic>=2.9.2
beautifulsoup4>=4.12.3
